Refactor RegistrationApp, make no functional changes to anything other than Mexico
