+++ api/src/test/java/org/openmrs/module/mirebalais/apploader/apps/PatientRegistrationAppTest.java
@@ -6,6 +6,7 @@ import org.openmrs.module.appframework.domain.AppDescriptor;
 import org.openmrs.module.haiticore.metadata.HaitiPersonAttributeTypes;
 import org.openmrs.module.metadatadeploy.api.MetadataDeployService;
 import org.openmrs.module.mirebalais.apploader.CustomAppLoaderConstants;
+import org.openmrs.module.mirebalais.apploader.apps.patientregistration.PatientRegistrationApp;
 import org.openmrs.module.pihcore.config.Config;
 import org.openmrs.module.pihcore.deploy.bundle.core.EncounterRoleBundle;
 import org.openmrs.module.pihcore.deploy.bundle.core.concept.SocioEconomicConcepts;
